Init component helpers; Move getComponentType there
