+++ src/scripts/react/common/componentHelpers.js
@@ -1,6 +1,10 @@
-export default (type) => {
+const getComponentType = (type) => {
   if (['extractor', 'writer', 'application'].includes(type)) {
     return type;
   }
   return 'component';
+};
+
+export {
+  getComponentType
 };
